[Fix] Fix assertion failure in generating system files.
Summary:
We use KindOfInt64 for system constants and an assertion failed when
KindOfInt32 was used for two new constants: OPENSSL_RAW_DATA and
OPENSSL_ZERO_PADDING. I made two changes:
(1) Change the type of these two constants to Int64
(2) Better error message instead of assertion failure

Test Plan:
make fast_tests
make slow_tests
The assertion failure would be now replaced by:
Exception: During code gen, constant with type Int32 is not expected\n

Reviewed By: mwilliams
Reviewers: mwilliams, macvicar
CC: hphp-diffs@lists, ps, mwilliams
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 255698

+++ src/idl/openssl.idl.php
@@ -30,13 +30,13 @@ CPP
 DefineConstant(
   array(
     'name'   => "OPENSSL_RAW_DATA",
-    'type'   => Int32,
+    'type'   => Int64,
   ));
 
 DefineConstant(
   array(
     'name'   => "OPENSSL_ZERO_PADDING",
-    'type'   => Int32,
+    'type'   => Int64,
   ));
